fix правки 2 app_01 HW6.2
