+++ .gitignore
@@ -9,7 +9,7 @@
 /.idea/assetWizardSettings.xml
 .DS_Store
 /build
-/.idea
+/.idea/
 /captures
 .externalNativeBuild
 .cxx
